requirements.txt
langchain==0.2.0
langchain-core==0.2.0
langchain-community==0.2.0
langchain-text-splitters==0.2.0
langsmith==0.1.17

faiss-cpu==1.7.4
sentence-transformers==2.7.0
transformers==4.41.1
torch==2.2.2

pypdf==4.2.0
scikit-learn==1.5.0

streamlit==1.35.0

google-generativeai==0.5.4
groq==0.9.0
httpx==0.27.0
python-dotenv==1.0.1